Move getConfig to utils
